+++ webapp/src/app/order/cart/page.tsx
@@ -1,0 +1,5 @@
+import OrderCartPage from "@/features/order-cart/OrderCartPage";
+
+export default function Page() {
+  return <OrderCartPage />;
+}
