--- webapp/src/app/order/page.tsx
@@ -1,5 +1,0 @@
-import OrderPage from "@/features/order/OrderPage";
-
-export default function Page() {
-  return <OrderPage />;
-}
